requirements.txt
boto3
mom
pycryptodome
argon2-cffi